refactor(codemods): add simple shared string prop migration util
