+++ packages/components/codemods/migrateBrandMomentMoodToVariant/transformBrandMomentMoodToVariant.ts
@@ -1,11 +1,12 @@
-import ts from "typescript"
-import { updateJsxElementWithNewProps } from "../utils"
-import { getPropValueText } from "../utils/getPropValueText"
+import { BrandMomentProps } from "~components/BrandMoment"
+import { migrateStringProp } from "../utils/migrateStringProp"
 
 const OLD_PROP_NAME = "mood"
 const NEW_PROP_NAME = "variant"
 
-const getNewVariantValue = (oldValue: string | undefined): string => {
+const getNewVariantValue = (
+  oldValue: Exclude<BrandMomentProps[typeof OLD_PROP_NAME], undefined>
+): Exclude<BrandMomentProps[typeof NEW_PROP_NAME], undefined> => {
   switch (oldValue) {
     case "informative":
       return "informative"
@@ -18,37 +19,8 @@ const getNewVariantValue = (oldValue: string | undefined): string => {
   }
 }
 
-/** Recurses through AST to find and update any jsx element that matched the tagName */
-export const transformBrandMomentMoodToVariant =
-  (context: ts.TransformationContext, tagName: string) =>
-  (rootNode: ts.Node): ts.Node => {
-    const visit = (node: ts.Node): ts.Node => {
-      if (ts.isJsxOpeningElement(node) || ts.isJsxSelfClosingElement(node)) {
-        if (node.tagName.getText() === tagName) {
-          const newAttributes = node.attributes.properties.map(attr => {
-            if (
-              ts.isJsxAttribute(attr) &&
-              attr.name.getText() === OLD_PROP_NAME
-            ) {
-              const oldValue =
-                attr.initializer && getPropValueText(attr.initializer)
-
-              if (oldValue) {
-                const newValue = getNewVariantValue(oldValue)
-                return ts.factory.createJsxAttribute(
-                  ts.factory.createIdentifier(NEW_PROP_NAME),
-                  ts.factory.createStringLiteral(newValue)
-                )
-              }
-            }
-
-            return attr
-          })
-
-          return updateJsxElementWithNewProps(node, newAttributes)
-        }
-      }
-      return ts.visitEachChild(node, visit, context)
-    }
-    return ts.visitNode(rootNode, visit)
-  }
+export const transformBrandMomentMoodToVariant = migrateStringProp(
+  OLD_PROP_NAME,
+  NEW_PROP_NAME,
+  getNewVariantValue
+)
