+++ packages/components/tsconfig.json
@@ -13,6 +13,6 @@
     "types": ["@testing-library/jest-dom"]
   },
   "files": ["types.d.ts"],
-  "include": ["src/**/*", "__tests__/**/*"],
+  "include": ["src/**/*", "__tests__/**/*", "codemods/**/*"],
   "exclude": ["node_modules"]
 }
